git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@15924 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ src/java/org/xmlBlaster/protocol/http/ajax/AjaxServlet.java
@@ -271,7 +271,7 @@ public class AjaxServlet extends HttpServlet {
 		}
 		String tmp = props.getProperty("maxInactiveInterval");
 		if (tmp != null)
-			this.maxInactiveInterval = Integer.valueOf(tmp);
+			this.maxInactiveInterval = Integer.valueOf(tmp).intValue();
 		this.blasterInstanceMap = new HashMap/*<String, BlasterInstance>*/();
 	}
